requirements.txt
Django==2.0
djangorestframework==3.7.3
pkg-resources==0.0.0
pytz==2017.3
